7, products.php,http://pm.5psolutions.net/public/index.php?path_info=projects/training-andrei-bontu/tasks/7, Implementing check's for the delete part

+++ public/products.php
@@ -15,7 +15,7 @@ $stmt->execute();
 $products = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
 //Delete products
-if (isset($_POST['product_id_to_remove'])) {
+if (isset($_POST['product_id_to_remove']) && $_POST['product_id_to_remove'] && !empty($_POST)) {
     $id = (int) $_POST['product_id_to_remove'];
 
     $stmt = $pdo->prepare('DELETE FROM products WHERE id = ?');
